+++ examples/table.rs
@@ -62,7 +62,7 @@ fn main() -> anyhow::Result<()> {
     // The first argument is the table index and the next 2 are the 2 arguments
     // to be passed to the function found in the table.
     let call_via_table: TypedFunction<(i32, i32, i32), i32> =
-        instance.exports.get_native_function("call_callback")?;
+        instance.exports.get_typed_function("call_callback")?;
 
     // And then call it with table index 1 and arguments 2 and 7.
     let result = call_via_table.call(1, 2, 7)?;